Events Overhaul
Updated the home page to get all the events from a component called events. This pulls from the data and then displays them based off of that. The link that is on each of the event cards takes you to pages of that specific event with the information pulled from the data.