Reset Password
The user can now reset their password. It creates a token that is sent in the email and creates a token in the database that stores the token string, expiration date (in two hours from creation), and the email it was sent to. This ensures that someone does not edit the account of someone else by making sure that the token and the user email match.